requirements.txt
resend==0.7.2
python-dotenv
pydantic

